requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1 